pyproject.toml
[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[project]
name = "life-optimizer"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0",
    "aiosqlite>=0.20",
    "Pillow>=10.0",
    "pyobjc-framework-Cocoa>=10.0",
    "anthropic>=0.49",
    "httpx>=0.27",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "jinja2>=3.1",
    "chromadb>=0.5",
    "sse-starlette>=2.0",
]

[project.scripts]
life-optimizer = "life_optimizer.cli:main"

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.24"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
